refactor into 3 packages
