+++ packages/contented-pipeline-md/src/plugins/Plugin.ts
@@ -1,0 +1,14 @@
+import { FileContentHeadings, Pipeline } from '@birthdayresearch/contented-pipeline';
+
+/**
+ * UnifiedData passed to unified plugins.
+ */
+export interface UnifiedContented {
+  pipeline: Pipeline;
+  headings: FileContentHeadings[];
+  fields: { [key: string]: any };
+  errors: {
+    type: string;
+    reason: string;
+  }[];
+}
