+++ packages/contented-pipeline-md/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["**/*.unit.ts"]
+}
