+++ packages/contented-processor/src/index.ts
@@ -1,3 +1,3 @@
 export * from './ContentedCodegen.js';
-export * from './ContentedPipeline.js';
 export * from './ContentedProcessor.js';
+export * from '@birthdayresearch/contented-pipeline';
